Create start and stop methods
The start method creates the primary thread on which the game will run.
This frees up the initial thread to run code in parallel, such as if
there was a need for a console window.

The stop method shuts down the primary thread and closes the window.
This method is most likely coded wrongly/inefficiently and should be
rewritten at some point soon.