Add new constructor and new method
The new constructor transfers all of the objects in the data variable
from the list of DataContainers that is passed to it and puts them in
its data variable.

The new method getMatchData returns a list of whatever class is given to
it containing all of the objects in the data variable that are an
instance of that class.