Make the orbit and body line up exactly
Previously, a body and its orbit did not line up when zoomed in close
enough. This was most likely caused because a circle drawn using the
Ellips2D class does not draw an exact circle as dictated by
trigonometry.

This was fixed by changing the orbit to a arc that starts at the angel
that the planet is at. The planet is then rendered at the starting point
of the arc.