Change orbit to use the parent rendered position
Previously used the true x and y coordinates in order to position the
orbit, this was changed to use the position that the parent is rendered
at by getting the parent's corresponding BodyUi class and getting the
center from that. This will prevent moons from being offset from the
parent body.